golang structure + basics (#4)
* chore: Set up minimal Go backend and frontend

This commit sets up the minimal structure for a Go backend and frontend. It includes the necessary files and directories to get started with development. The backend includes a basic server setup, while the frontend includes a simple HTML file and a CSS file generated from Tailwind CSS.

* chore(server): add middleware for logging and recovery

Add gin.Logger() and gin.Recovery() middleware to the server to handle logging and recover from panics.

* fix(.gitignore): add frontend/static/style.css to gitignore

The file frontend/static/style.css was added to the gitignore file to prevent it from being tracked by git.

* feat: add WebAssembly support

This commit adds support for WebAssembly by including the necessary JavaScript code, `wasm_exec.js`, and making changes to the `index.html` file to load and run the WebAssembly module.

* chore: move to single package

Move all code to a single package for better organization and maintainability.

* feat(frontend): improve styling of CelestralSrc

- Updated the font tracking for the main heading to make it wider and more visually appealing.
- Modified the click action of the div to use a POST request instead of a GET request.

* chore(justfile): update watch command in justfile

Update the watch command in the justfile to use the new "c" flag for continuous mode. This ensures that the backend is continuously run when there are file changes in the frontend static files or any go files.

* feat(server): add custom logger

This commit adds a custom logger to the server code in order to customize the log output. The custom logger uses the `gin.LoggerWithConfig` function to configure the logger with a custom log formatter. The log formatter is defined in the `logger` package and uses the `lipgloss` library to style the log output. The log messages now include the timestamp, status code, request latency, response size, request method, client IP, and request path.

* feat(main): add canvas animation

Add canvas animation to the main function to create a starry sky effect.

* feat(server): add gzip compression middleware

Add the `gin-contrib/gzip` package to enable gzip compression middleware in the server. This will improve the performance of the server by compressing the response body before sending it to the client.

* chore(justfile): Add chrome dev tools commands

Add commands to start chrome with developer port and reload the chrome webpage for easier debugging.

* feat: add concept.png

Add a new concept.png file to the codebase.